requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow~=8.2.0

# logging -------------------------------------
# wandb

# plotting ------------------------------------

# export --------------------------------------
# coremltools==4.0
# onnx>=1.8.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------

tensorflow~=2.5.0rc0
Flask~=2.0.1
PyQt5~=5.15.4
pip~=21.1.2
wheel~=0.36.2
lxml~=4.6.3
PyYAML~=5.4.1
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
Jinja2~=3.0.1
Click~=7.0
itsdangerous~=2.0.1
Cython~=0.29.22
setuptools~=57.0.0
pytz~=2019.2
nose~=1.3.7
six~=1.15.0
python-dateutil~=2.8.0
cycler~=0.10.0
pyparsing~=2.4.2
kiwisolver~=1.3.1
Keras~=2.4.3